move staking page contents into components.rs
